(perl #127663) ensure abandoned work files are cleaned up
If the user code that's processing in-place edit files exits or dies,
previously this would leave the in-place edit work file in place.

This patch deletes the temp file, abandoning the edit, which is a change
from the old in-place editing, so maybe this should do the normal
success processing currently done in do_close().