(perl #127663) add our own mkstemp() implementation
this needs a couple of adjustments:

- needs a wrapper that calls either the system mkstemp() with a fallback
to Perl_my_mkstemp().

Not based on any particular implementation, the BSD implementations
tend to be wrappers around a megafunction that also does a few variations
of mkstemp() and mkdtemp(), which we don't need (yet.)

One implementation I found, part of the heimdal crypto library, was
simpler, but horrible.